requirements.txt
allure-python-commons~=2.13.5
pytest~=8.3.5
requests~=2.32.3